+++ src/ModuleInterface.php
@@ -6,7 +6,6 @@ use Aura\Di\ContainerConfigInterface;
 interface ModuleInterface extends ContainerConfigInterface
 {
     public function require();
-    public function requireDev();
     public function conflict();
     public function replace();
 }