+++ src/ModuleLoaderInterface.php
@@ -5,6 +5,5 @@ use Aura\Di\ContainerConfigInterface;
 
 interface ModuleLoaderInterface extends ContainerConfigInterface
 {
-    public function isDev();
     public function loaded($name);
 }