+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 0.4.0 - 2017-02-17
+Transitioned from `isDev` and `requireDev` to `environment` and `require{Environment}`
+### Changed
+- `$isDev = false` in ModuleLoader constructor to `$environment = ''`
+### Removed
+- `isDev()` method from `ModuleLoader` and `ModuleLoaderInterface`
+- `requireDev()` method from `Module` and `ModuleInterface`
+### Added
+- Support for new `require{Environment}` methods (ex: `requireSpecialEnvironment`)
+
 ## 0.3.0 - 2016-11-30
 Bringing up to speed with new Cadre.Package skeleton
 ### Added